finfished

+++ src/components/productCard.vue
@@ -32,7 +32,33 @@ export default {
       this.SET_EDITPRODUCT(this.currentProd)
     },
     deleteProd () {
-      this.deleteProduct(this.id)
+      this.$swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+      }).then((result) => {
+        if (result.value) {
+          this.deleteProduct(this.id)
+          .then(result => {
+            this.$swal.fire(
+              'Deleted!',
+              `${result.result.data.message}`,
+              'success'
+            )
+          })
+          .catch(err => {
+            this.$swal.fire({
+            icon: 'error',
+            title: 'Oops...',
+            text: 'Something went wrong!',
+          })
+          })
+        }
+      })
     }
   },
   computed: {
